composerの標準的な名前空間を導入するため、ディレクトリの整理
