--- lib/Omise.php
@@ -1,12 +1,0 @@
-<?php
-
-define('OMISE_PHP_LIB_VERSION', '1.0.0');
-define('OMISE_API_VERSION', '2014-07-27');
-
-require_once dirname(__FILE__).'/omise/OmiseAccount.php';
-require_once dirname(__FILE__).'/omise/OmiseBalance.php';
-require_once dirname(__FILE__).'/omise/OmiseToken.php';
-require_once dirname(__FILE__).'/omise/OmiseCharge.php';
-require_once dirname(__FILE__).'/omise/OmiseCustomer.php';
-require_once dirname(__FILE__).'/omise/OmiseTransfer.php';
-require_once dirname(__FILE__).'/omise/OmiseTransaction.php';
