+++ lib/Omise/OmisePHP/Omise.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Omise\OmisePHP;
+
+define('OMISE_PHP_LIB_VERSION', '1.0.0');
+define('OMISE_API_VERSION', '2014-07-27');
+
+require_once dirname(__FILE__).'/OmiseAccount.php';
+require_once dirname(__FILE__).'/OmiseBalance.php';
+require_once dirname(__FILE__).'/OmiseToken.php';
+require_once dirname(__FILE__).'/OmiseCharge.php';
+require_once dirname(__FILE__).'/OmiseCustomer.php';
+require_once dirname(__FILE__).'/OmiseTransfer.php';
+require_once dirname(__FILE__).'/OmiseTransaction.php';
