+++ lib/Omise/OmisePHP/OmiseBalance.php
@@ -1,15 +1,17 @@
 <?php
 
+namespace Omise\OmisePHP;
+
 require_once dirname(__FILE__).'/res/OmiseApiResourceSingleton.php';
 
-class OmiseAccount extends OmiseApiResourceSingleton {
-  const ENDPOINT = 'account';
+class OmiseBalance extends OmiseApiResourceSingleton {
+  const ENDPOINT = 'balance';
 
   /**
-   * Retrieves an account.
+   * Retrieves a current balance in the account.
    * @param string $publickey
    * @param string $secretkey
-   * @return OmiseAccount
+   * @return OmiseBalance
    */
   public static function retrieve($publickey = null, $secretkey = null) {
     return parent::retrieve(get_class(), self::getUrl(), $publickey, $secretkey);
@@ -24,7 +26,6 @@ class OmiseAccount extends OmiseApiResourceSingleton {
   }
 
   /**
-   *
    * @param string $cardID
    * @return string
    */
