+++ lib/Omise/OmisePHP/OmiseCard.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Omise\OmisePHP;
+
 require_once dirname(__FILE__).'/res/OmiseApiResource.php';
 require_once dirname(__FILE__).'/OmiseCustomer.php';
