+++ lib/Omise/OmisePHP/OmiseCustomer.php
@@ -1,27 +1,30 @@
 <?php
 
-require_once dirname(__FILE__).'/res/OmiseApiResource.php';
+namespace Omise\OmisePHP;
 
-class OmiseTransfer extends OmiseApiResource {
-  const ENDPOINT = 'transfers';
+require_once dirname(__FILE__).'/res/OmiseApiResource.php';
+require_once dirname(__FILE__).'/res/obj/OmiseCardList.php';
+
+class OmiseCustomer extends OmiseApiResource {
+  const ENDPOINT = 'customers';
 
   /**
-   * Retrieves a transfer.
+   * Retrieves a customer.
    * @param string $id
    * @param string $publickey
    * @param string $secretkey
-   * @return OmiseTransfer
+   * @return OmiseCustomer
    */
   public static function retrieve($id = '', $publickey = null, $secretkey = null) {
     return parent::retrieve(get_class(), self::getUrl($id), $publickey, $secretkey);
   }
 
   /**
-   * Creates a transfer.
-   * @param unknown $params
+   * Creates a new customer.
+   * @param array $params
    * @param string $publickey
    * @param string $secretkey
-   * @return OmiseTransfer
+   * @return OmiseCustomer
    */
   public static function create($params, $publickey = null, $secretkey = null) {
     return parent::create(get_class(), self::getUrl(), $params, $publickey, $secretkey);
@@ -32,7 +35,7 @@ class OmiseTransfer extends OmiseApiResource {
    * @see OmiseApiResource::reload()
    */
   public function reload() {
-    if($this['object'] === 'transfers') {
+    if($this['object'] === 'customer') {
       parent::reload(self::getUrl($this['id']));
     } else {
       parent::reload(self::getUrl());
@@ -40,17 +43,10 @@ class OmiseTransfer extends OmiseApiResource {
   }
 
   /**
-   * Updates the transfer amount.
-   */
-  public function save() {
-    $this->update(array('amount' => $this['amount']));
-  }
-
-  /**
    * (non-PHPdoc)
    * @see OmiseApiResource::update()
    */
-  protected function update($params) {
+  public function update($params) {
     parent::update(self::getUrl($this['id']), $params);
   }
 
@@ -68,6 +64,16 @@ class OmiseTransfer extends OmiseApiResource {
    */
   public function isDestroyed() {
     return parent::isDestroyed();
+  }
+
+  /**
+   * Gets a list of all cards belongs to this customer.
+   * @return OmiseCardList
+   */
+  public function getCards() {
+    if($this['object'] === 'customer') {
+      return new OmiseCardList($this, $this->_publickey, $this->_secretkey);
+    }
   }
 
   /**
