+++ lib/Omise/OmisePHP/exception/OmiseExceptions.php
