+++ lib/Omise/OmisePHP/res/OmiseApiResourceSingleton.php
